Completed Part3, addressed comments

+++ src/Transactions.cs
@@ -1,9 +1,3 @@
-using System;
-using System.Globalization;
-using System.IO;
-using System.Transactions;
-using CsvHelper;
-using CsvHelper.Configuration;
 
 namespace SupportBank {
     public class Transaction
@@ -13,7 +7,7 @@ namespace SupportBank {
             this.From = from;
             this.To = to;
             this.Narrative = narrative;
-            this.Amount = (int)float.Parse(amount)*100;
+            this.Amount = (int)(double.Parse(amount)*100);
         }
         public string Date { get; }
         public string From { get; }
